Фиксы багов на сервере и в App, Добалено разделение очереди на вопросы в инспекции и менеджер задач

+++ ApiMobaileServise/BackgraundService/Queue/QueueWorkerAsk.cs
@@ -2,16 +2,16 @@
 using FluentScheduler;
 using System;
 using System.Collections.Generic;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace ApiMobaileServise.BackgraundService.Queue
 {
-    public class QueueWorker : IJob
+    public class QueueWorkerAsk : IJob
     {
         private ManagerMobileApi managerMobileApi = null;
         public static List<string> queues = new List<string>();
         public static int countQueues = 0;
+        public static bool isWork = false;
 
         public void Execute()
         {
@@ -21,22 +21,20 @@ namespace ApiMobaileServise.BackgraundService.Queue
 
         private async void QWork()
         {
-            if (queues.Count != 0)
+            if (queues.Count != 0 && !isWork)
             {
+                isWork = true;
                 int tmpCount = queues.Count > 5000 ? queues.Count - (countQueues - 5000) : queues.Count;
                 for (int i = 0; i < tmpCount; i++)
                 {
-                    if(queues[i].Split(',')[0] == "Load")
+                    if (queues[i].Split("&,&")[0] == "SaveAnsver")
                     {
-                        await managerMobileApi.LoadTask(queues[i].Split(',')[1], queues[i].Split(',')[2]);
-                    }
-                    else if (queues[i].Split(',')[0] == "End")
-                    {
-                        await managerMobileApi.EndTask(queues[i].Split(',')[1], queues[i].Split(',')[2]);
+                        await managerMobileApi.SaveAsk(queues[i].Split("&,&")[1], Convert.ToInt32(queues[i].Split("&,&")[2]), queues[i].Split("&,&")[3]);
                     }
                     countQueues--;
                 }
                 queues.RemoveRange(0, tmpCount);
+                isWork = false;
             }
         }
     }
